Move admin/datamodel/metadata.js to admin/datamodel/datamodel.js
